Added section on filesystem security hardening
Added a new section that goes into filesystem security hardening for Fedora 43 workstations that will be used for code development. Also, adjusted some language here and there, and linked the JetBrains section to another guide in this repo that walks the user through setting up a server for remote builds and code execution.